+++ core/deltachat-ffi/src/lib.rs
@@ -493,7 +493,6 @@ pub unsafe extern "C" fn dc_event_get_id(event: *mut dc_event_t) -> libc::c_int 
         EventType::ImapConnected(_) => 102,
         EventType::SmtpMessageSent(_) => 103,
         EventType::ImapMessageDeleted(_) => 104,
-        EventType::ImapMessageMoved(_) => 105,
         EventType::ImapInboxIdle => 106,
         EventType::NewBlobFile(_) => 150,
         EventType::DeletedBlobFile(_) => 151,
@@ -560,7 +559,6 @@ pub unsafe extern "C" fn dc_event_get_data1_int(event: *mut dc_event_t) -> libc:
         | EventType::ImapConnected(_)
         | EventType::SmtpMessageSent(_)
         | EventType::ImapMessageDeleted(_)
-        | EventType::ImapMessageMoved(_)
         | EventType::ImapInboxIdle
         | EventType::NewBlobFile(_)
         | EventType::DeletedBlobFile(_)
@@ -635,7 +633,6 @@ pub unsafe extern "C" fn dc_event_get_data2_int(event: *mut dc_event_t) -> libc:
         | EventType::ImapConnected(_)
         | EventType::SmtpMessageSent(_)
         | EventType::ImapMessageDeleted(_)
-        | EventType::ImapMessageMoved(_)
         | EventType::ImapInboxIdle
         | EventType::NewBlobFile(_)
         | EventType::DeletedBlobFile(_)
@@ -730,7 +727,6 @@ pub unsafe extern "C" fn dc_event_get_data2_str(event: *mut dc_event_t) -> *mut 
         | EventType::ImapConnected(msg)
         | EventType::SmtpMessageSent(msg)
         | EventType::ImapMessageDeleted(msg)
-        | EventType::ImapMessageMoved(msg)
         | EventType::NewBlobFile(msg)
         | EventType::DeletedBlobFile(msg)
         | EventType::Warning(msg)