+++ core/deltachat-rpc-client/src/deltachat_rpc_client/const.py
@@ -38,7 +38,6 @@ class EventType(str, Enum):
     IMAP_CONNECTED = "ImapConnected"
     SMTP_MESSAGE_SENT = "SmtpMessageSent"
     IMAP_MESSAGE_DELETED = "ImapMessageDeleted"
-    IMAP_MESSAGE_MOVED = "ImapMessageMoved"
     IMAP_INBOX_IDLE = "ImapInboxIdle"
     NEW_BLOB_FILE = "NewBlobFile"
     DELETED_BLOB_FILE = "DeletedBlobFile"