+++ core/src/imap/client.rs
@@ -78,7 +78,6 @@ pub(crate) async fn identify_server(
     };
     let capabilities = Capabilities {
         can_idle: caps.has_str("IDLE"),
-        can_move: caps.has_str("MOVE"),
         can_check_quota: caps.has_str("QUOTA"),
         can_metadata: caps.has_str("METADATA"),
         can_compress: caps.has_str("COMPRESS=DEFLATE"),